Added list of currently supported features.
Most of these I wrote years ago, so some of it, like the Elgg authentication support, may be quite out of date. But I managed to tweak it so it works with current PHP so I can use it as the base system for this project. It was intended for something similar anyway.